Add Manage Data widget
The widget makes it possible to import activities with control over the time frame. It also has a basic way to set conversation properties from a given yaml (wip)

+++ src/graphql/schemas.js
@@ -178,7 +178,7 @@ const typeDefs = gql`
     text: String
     textHtml: String
     timestamp: String
-    timestampInt: Int
+    timestampInt: Float
     url: String
     project: Project! @relationship(type: "OWNS", direction: IN)
     member: Member! @relationship(type: "DID", direction: IN)
